requirements.txt
# Core dependencies
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0
pyyaml>=6.0
matplotlib>=3.7.0

# ONNX export
onnx>=1.14.0
#onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0

# Development
tqdm>=4.65.0
pandas>=2.0.0